refactor: layout-> layoutComponent + layoutContainer, app->AppComponent+AppContainer
