+++ src/app/AppComponent.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import LayoutContainer from 'src/widgets/layout/LayoutContainer';
+
+const AppComponent: React.FC = () => {
+  return <LayoutContainer />;
+};
+
+export default AppComponent;
