+++ src/widgets/modal/Modal.tsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import styles from './modal.module.css';
 import { useTranslation } from 'react-i18next';
 import { createPortal } from 'react-dom';
